+++ src/KMonad/Args/Joiner.hs
@@ -25,6 +25,7 @@ module KMonad.Args.Joiner
 where
 
 import KMonad.Prelude hiding (uncons)
+import KMonad.Util
 
 import KMonad.Args.Types
 
@@ -80,6 +81,7 @@ data JoinError
   | NestedTrans
   | InvalidComposeKey
   | LengthMismatch   Text Int Int
+  | NotAKeycode      DefButton
 
 instance Show JoinError where
   show e = case e of
@@ -110,6 +112,7 @@ instance Show JoinError where
       [ "Mismatch between length of 'defsrc' and deflayer <", T.unpack t, ">\n"
       , "Source length: ", show s, "\n"
       , "Layer length: ", show l ]
+    NotAKeycode       b   -> "Encountered non keycode button in 'defsrc': " <> show b
 
 
 instance Exception JoinError
@@ -351,11 +354,16 @@ joinAliases als = do
 
   M.traverseWithKey =<< go $ M.fromList als'
  where
-  go mp t (KRef t') = case (compare t t', M.lookup t' mp) of
-      (EQ, _) -> throwError $ CyclicAlias t
-      (_, Nothing) -> throwError $ MissingAlias t'
-      (_, Just b) -> go mp t b
+  go mp t b@(KRef t')
+    | t == t' = throwError $ CyclicAlias t
+    | otherwise = withDerefAlias mp (go mp t) b
   go _ _ b = pure b
+
+withDerefAlias :: Aliases -> (DefButton -> J a) -> DefButton -> J a
+withDerefAlias als f (KRef t) = case M.lookup t als of
+  Just b -> f b
+  Nothing -> throwError $ MissingAlias t
+withDerefAlias _ f b = f b
 
 --------------------------------------------------------------------------------
 -- $but
@@ -383,9 +391,7 @@ joinButton ns als =
       isps l = traverse go . maybe l ((`intersperse` l) . KPause . fi)
   in \case
     -- Variable dereference
-    KRef t -> case M.lookup t als of
-      Nothing -> throwError $ MissingAlias t
-      Just b  -> joinButton ns als b
+    b@(KRef _) -> withDerefAlias als (joinButton ns als) b
 
     -- Various simple buttons
     KEmit c -> ret $ emitB c
@@ -451,20 +457,23 @@ joinButton ns als =
 --------------------------------------------------------------------------------
 -- $src
 
-type Sources = M.HashMap (Maybe Text) DefSrc
+type Sources = M.HashMap (Maybe Text) [Keycode]
 
 -- | Build up a hashmap of text to source mappings.
-joinSources :: [DefSrc] -> J Sources
-joinSources = foldM joiner mempty
+joinSources :: Aliases -> [DefSrc] -> J Sources
+joinSources als = foldM joiner mempty
   where
    joiner :: Sources -> DefSrc -> J Sources
-   joiner sources src@DefSrc{ _srcName = n, _keycodes = ks }
+   joiner sources DefSrc{ _srcName = n, _keycodes = ks }
      | n `M.member` sources = throwError $ DuplicateSource n
-     | not (null dups)      = throwError $ DuplicateKeyInSource n dups
-     | otherwise            = pure $ M.insert n src sources
+     | otherwise            = do
+        ks' <- for ks $ withDerefAlias als joinKeycode
+        let dups = NE.head <$> duplicatesWith id ks'
+        unless (null dups) $ throwError $ DuplicateKeyInSource n dups
+        pure $ M.insert n ks' sources
     where
-     dups :: [Keycode]
-     dups = concatMap (take 1) . filter ((> 1) . length) . group . sort $ ks
+     joinKeycode (KEmit k) = pure k
+     joinKeycode b         = throwError $ NotAKeycode b
 
 --------------------------------------------------------------------------------
 -- $kmap
@@ -478,7 +487,7 @@ joinKeymap srcs als lys = do
   let f acc x = if x `elem` acc then throwError $ DuplicateLayer x else pure (x:acc)
   nms   <- foldM f [] $ map _layerName lys     -- Extract all names
   als'  <- joinAliases als                     -- Join aliases into 1 hashmap
-  srcs' <- joinSources  srcs                   -- Join all sources into 1 hashmap
+  srcs' <- joinSources als' srcs               -- Join all sources into 1 hashmap
   lys'  <- mapM (joinLayer als' nms srcs') lys -- Join all layers
   -- Return the layerstack and the name of the first layer
   pure (L.mkLayerStack lys', _layerName . fromJust . headMaybe $ lys)
@@ -496,7 +505,7 @@ joinLayer als ns srcs l@(DefLayer n settings) = do
   implAround <- getImplAround l
 
   src <- case M.lookup assocSrc srcs of
-    Just src -> pure $ src^.keycodes
+    Just src -> pure src
     Nothing  -> throwError $ MissingSource assocSrc
   -- Ensure length-match between src and buttons
   when (length bs /= length src) $