+++ test/KMonad/KeycodeSpec.hs
@@ -3,12 +3,11 @@ module KMonad.KeycodeSpec (spec) where
 import KMonad.Keyboard.Keycode
 import KMonad.Util.MultiMap as Q
 import KMonad.Prelude
+import KMonad.Util
 import RIO.List (sort)
 
 import qualified KMonad.Keyboard.IO.Mac.Types as Mac (kcMapRaw)
 import qualified KMonad.Keyboard.IO.Windows.Types as Win (winCodeKeyCodeMapping)
-
-import qualified RIO.NonEmpty as N
 
 import Test.Hspec
 
@@ -16,7 +15,7 @@ spec :: Spec
 spec = do
 
   it "No duplicate keycode names" $
-    dupsWith snd (keyNames ^.. Q.itemed) `shouldBe` []
+    duplicatesWith snd (keyNames ^.. Q.itemed) `shouldBe` []
 
   describe "MacOS keycodes" $ checkOsMapping Mac.kcMapRaw
   describe "Windows keycodes" $ checkOsMapping Win.winCodeKeyCodeMapping
@@ -29,6 +28,3 @@ checkOsMapping m = do
   let sortedM' = snd <$> difference
   it "Raw list is sorted" $
     m' `shouldBe` sortedM' -- Only shows unsorted indexes when failing
-
-dupsWith :: Ord b => (a -> b) -> [a] -> [NonEmpty a]
-dupsWith f = filter (not . null . N.tail) . N.groupAllWith f