+++ changelog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
 - Update Karabiner-DriverKit to 5.0.0 (#937)
   You will now need to [start the server yourself](doc/installation.md#starting-the-dext-daemon).
 - Aliases can now refer to later aliases instead of just earlier ones. (#992)
+- Aliases can now be used in `defsrc` if they are a keycode (#992)
 
 ### Fixed